fix: controllers functionality (CRUD methods) fixed to handle errors
The erros are handled by using try/catch and throwing Errors to manipulate the json response

+++ src/application/useCases/Course/courseController.ts
@@ -1,61 +1,92 @@
 import { Course } from "../../../domain/entities/course";
-import { CourseRepository } from "../../repositories/IcoursesRepository";
+import {
+  CourseRepository,
+  CourseSupaBaseResponse,
+} from "../../repositories/IcoursesRepository";
+
 import { supabase } from "../../../utils/connect_db";
-import { ReadStudentUseCase } from "../Student/readStudentUseCase";
-import { StudentController } from "../Student/studentController";
 
 export class CourseController implements CourseRepository {
   // Save functionality
-  async save(course: Course): Promise<Object> {
-    const result = await supabase.from("courses").insert([
-      {
-        id: course.id,
-        title_course: course.props.title_course,
-        description: course.props.description,
-        time_course: course.props.time_course,
-      },
-    ]);
+  async save(course: Course): Promise<CourseSupaBaseResponse> {
+    if (course.props.title_course.length !== 0) {
+      const result = await supabase.from("courses").insert([
+        {
+          id: course.id,
+          title_course: course.props.title_course.toLocaleLowerCase(),
+          description: course.props.description,
+          time_course: course.props.time_course,
+        },
+      ]);
 
-    return result;
+      if (result.error) {
+        throw new Error(
+          `[ERROR]: ${result.error.code} | ${result.error.message} | details: ${result.error.details} | hint: ${result.error.hint}`
+        );
+      }
+
+      return { message: result };
+    }
+    throw new Error(
+      "Title Course need a valid string length (greater than 3 characters)"
+    );
   }
 
   // Read functionality
-  async read(id?: string): Promise<Object> {
-    if (id) {
+  async read(id?: string): Promise<CourseSupaBaseResponse> {
+    if (id && id.length !== 0) {
       const result = await supabase.from("courses").select().eq("id", id);
-      return result;
+
+      if (result.error) {
+        throw new Error("There's no course with this Identifier");
+      }
+
+      return { message: result };
     }
 
     const result = await supabase.from("courses").select();
 
-    return result;
+    if (result.error || result.data.length === 0) {
+      throw new Error("There's no course yet");
+    }
+
+    return { message: result };
   }
 
-  async update(id: string, newData: Course): Promise<Object> {
-    if (id && newData) {
+  async update(id: string, newData: Course): Promise<CourseSupaBaseResponse> {
+    if (id && newData && id.length !== 0) {
       const result = await supabase
         .from("courses")
         .update({
-          id: newData.id,
-          title_course: newData.props.title_course,
+          title_course: newData.props.title_course.toLocaleLowerCase(),
           description: newData.props.description,
           time_course: newData.props.time_course,
         })
         .match({ id: id });
 
-      return result;
+      if (result.error) {
+        throw new Error(
+          `[ERROR]: ${result.error.code} | ${result.error.message} | details: ${result.error.details} | hint: ${result.error.hint}`
+        );
+      }
+
+      return { message: result };
     }
 
-    return { error: "Something wrong ocurred" };
+    throw new Error("Verify the ID and newData");
   }
 
-  async delete(id: string): Promise<Object> {
-    if (id) {
+  async delete(id: string): Promise<CourseSupaBaseResponse> {
+    if (id && id.length !== 0) {
       const result = await supabase.from("courses").delete().match({ id: id });
 
-      return result;
+      if (result.error || result.data.length === 0) {
+        throw new Error("Impossible to delete. There's no course like that");
+      }
+
+      return { message: result, status: "DELETED" };
     }
 
-    return { error: "Impossible to delete. Verify the ID Course" };
+    throw new Error("Impossible to delete. Verify the ID Course");
   }
 }
